Change the link 2014 e 2015 Pnad data
Lines 184 and 185 need updating.

I'm trying to analyze the PNAD 2015, but I have several problems when I try to update the variables. 

In my case my computer is Mac and I also need to alter the 203 line: if( year %in% 2013:2015 & .Platform$OS.type != 'windows' ) Encoding( files ) <- 'UTF-8' else Encoding( files ) <- 'latin1' to 2013:2015. However, I'm afraid of causing a problem because when I tried to analyzed the pnad I don't get to update variables. I tried to use update from survey, but It was unsuccessful. The new variable is NA. So I altered the pnad2015 table and after stratify and poststratify, but the variable is not recognized in survey.